+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 #### Breaking Changes
 
 * remove support for last keydown code tracking on form controls (this feature has been superseded by user defined properties which allow associating any kind of metadata with a control)
+* rename `groupUpdateReducer` to `createFormGroupReducerWithUpdate` in order to make it clearer that the function itself is not a reducer
 
 #### Features